Move check on IS_SCIPY_INSTALLED to svir __init__.py
