+++ svir/recovery_modeling/building.py
@@ -4,7 +4,7 @@ from PyQt4.QtCore import QSettings
 from svir.utilities.shared import RECOVERY_DEFAULTS
 from svir.utilities.utils import get_layer_setting
 
-from svir.irmt import IS_SCIPY_INSTALLED
+from svir import IS_SCIPY_INSTALLED
 if IS_SCIPY_INSTALLED:
     from scipy.stats import norm
